requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.0.3
selenium==4.15.0
scrapy==2.11.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pytest==7.4.3
fake-useragent==1.4.0